requirements.txt
openrgb-python
flask
flask-restful
